Added Screen Blink when damaged, Zombie Kill Count, and Kill Streak Text Pop up
Added a feature where the screen will blink red when taken damage.
Added a zombie kill count text on the gui that keeps track of the number
of zombies killed
Added a Kill Streak text on the gui that pops up every 5 zombies kills.
This number and text can tweaked.